mv to_thread() to utils
